Further refactored to include all commune logic in cambodia.py
