+++ backend/utils/settings.py
@@ -1,13 +1,4 @@
-import geopandas as gpd
-import os
-from countries.cambodia import normalize_province_name  # Import normalization function
+# backend/utils/settings.py
+from countries.cambodia import get_communes_geodataframe  # Import from cambodia.py
 
-def get_communes_geodataframe():
-    """Load the GeoDataFrame for Cambodia's communes."""
-    geojson_file = os.path.join(os.getcwd(), "boundaries", "cambodia_communes.geojson")
-    try:
-        communes_gdf = gpd.read_file(geojson_file)
-        communes_gdf["normalized_NAME_1"] = communes_gdf["NAME_1"].apply(normalize_province_name)
-    except Exception as e:
-        raise RuntimeError(f"Error reading GeoJSON file: {str(e)}")
-    return communes_gdf
+# This file should now only include global settings or configurations.
