+++ backend/weather/temperature.py
@@ -1,3 +1,5 @@
+# backend/weather/temperature.py
+
 import ee
 import pandas as pd
 
@@ -11,45 +13,53 @@ def retrieve_temperature_data(province_gdf, start_date: str, end_date: str):
     """
     result_df = pd.DataFrame()
 
-    # Iterate through each commune
-    for _, commune in province_gdf.iterrows():
-        commune_name = commune["NAME_3"]
+    # Iterate through each district in the province
+    unique_districts = province_gdf["NAME_2"].unique()
+    for district in unique_districts:
+        print(f"[INFO] Processing district: {district}")  # Print statement to track progress
 
-        # Define the geometry
-        if commune["geometry"].geom_type == "MultiPolygon":
-            polygons = [ee.Geometry.Polygon(list(poly.exterior.coords)) for poly in commune["geometry"].geoms]
-            commune_geometry = ee.Geometry.MultiPolygon(polygons)
-        else:
-            commune_geometry = ee.Geometry.Polygon(commune["geometry"]["coordinates"])
+        # Filter for communes within the current district
+        district_communes = province_gdf[province_gdf["NAME_2"] == district]
 
-        # Fetch ERA5 Daily Temperature data for the commune's polygon over the specified time period
-        temperature = (
-            ee.ImageCollection("ECMWF/ERA5/DAILY")
-            .filterDate(start_date, end_date)
-            .select("mean_2m_air_temperature")
-        )
+        # Retrieve temperature data for each commune in the district
+        for _, commune in district_communes.iterrows():
+            commune_name = commune["NAME_3"]
 
-        # Add a 'date' property to each image
-        temperature = temperature.map(lambda image: image.set("date", image.date().format("YYYY-MM-dd")))
+            # Define the geometry
+            if commune["geometry"].geom_type == "MultiPolygon":
+                polygons = [ee.Geometry.Polygon(list(poly.exterior.coords)) for poly in commune["geometry"].geoms]
+                commune_geometry = ee.Geometry.MultiPolygon(polygons)
+            else:
+                commune_geometry = ee.Geometry.Polygon(commune["geometry"]["coordinates"])
 
-        # Extract the data for each date using `reduceRegion`
-        def extract_daily_data(image):
-            mean_temperature = image.reduceRegion(
-                reducer=ee.Reducer.mean(),
-                geometry=commune_geometry,
-                scale=5000,
-                maxPixels=1e13
+            # Fetch ERA5 Daily Temperature data for the commune's polygon over the specified time period
+            temperature = (
+                ee.ImageCollection("ECMWF/ERA5/DAILY")
+                .filterDate(start_date, end_date)
+                .select("mean_2m_air_temperature")
             )
-            return image.set("mean_temperature", mean_temperature.get("mean_2m_air_temperature"))
 
-        # Apply the `extract_daily_data` function
-        time_series = temperature.map(extract_daily_data).getInfo()
+            # Add a 'date' property to each image
+            temperature = temperature.map(lambda image: image.set("date", image.date().format("YYYY-MM-dd")))
 
-        # Convert to a DataFrame
-        daily_data = {entry["properties"]["date"]: entry["properties"]["mean_temperature"] for entry in time_series["features"] if "mean_temperature" in entry["properties"]}
-        daily_df = pd.DataFrame(list(daily_data.items()), columns=["Date", commune_name])
-        
-        # Merge results
-        result_df = pd.merge(result_df, daily_df, on="Date", how="outer") if not result_df.empty else daily_df
+            # Extract the data for each date using `reduceRegion`
+            def extract_daily_data(image):
+                mean_temperature = image.reduceRegion(
+                    reducer=ee.Reducer.mean(),
+                    geometry=commune_geometry,
+                    scale=5000,
+                    maxPixels=1e13
+                )
+                return image.set("mean_temperature", mean_temperature.get("mean_2m_air_temperature"))
+
+            # Apply the `extract_daily_data` function
+            time_series = temperature.map(extract_daily_data).getInfo()
+
+            # Convert to a DataFrame
+            daily_data = {entry["properties"]["date"]: entry["properties"]["mean_temperature"] for entry in time_series["features"] if "mean_temperature" in entry["properties"]}
+            daily_df = pd.DataFrame(list(daily_data.items()), columns=["Date", commune_name])
+            
+            # Merge results
+            result_df = pd.merge(result_df, daily_df, on="Date", how="outer") if not result_df.empty else daily_df
 
     return result_df
